feat: KeyboardInput.java
Added functionality to swtich to a running state and revert to idle
state if the key is released.

Commit by richiesuper

+++ src/main/Game.java
@@ -21,7 +21,7 @@ public class Game implements Runnable {
 			gamePanel.repaint();
 
 			try {
-				Thread.sleep(Constants.Game.FRAME_PERIOD);
+				Thread.sleep((long) Constants.Game.FRAME_PERIOD);
 			} catch (InterruptedException e) {
 				e.printStackTrace();
 			}
